bottles: more porting to generic layer.
