+++ Library/Homebrew/test/bottles/testball_bottle-0.1.macintosh_intel.bottle.tar.gz
@@ -1,0 +1,1 @@
+testball_bottle-0.1.yosemite.bottle.tar.gz
