+++ Library/Homebrew/test/test_bottle_collector.rb
@@ -30,23 +30,9 @@ class BottleCollectorTests < Homebrew::TestCase
     assert_nil checksum_for(:foo)
   end
 
-  def test_collector_finds_or_later_tags
-    @collector[:lion_or_later] = "foo"
-    assert_equal ["foo", :lion_or_later], checksum_for(:mountain_lion)
-    assert_nil checksum_for(:snow_leopard)
-  end
-
   def test_collector_prefers_exact_matches
     @collector[:lion_or_later] = "foo"
     @collector[:mountain_lion] = "bar"
     assert_equal ["bar", :mountain_lion], checksum_for(:mountain_lion)
-  end
-
-  def test_collector_finds_altivec_tags
-    @collector[:tiger_altivec] = "foo"
-    assert_equal ["foo", :tiger_altivec], checksum_for(:tiger_g4)
-    assert_equal ["foo", :tiger_altivec], checksum_for(:tiger_g4e)
-    assert_equal ["foo", :tiger_altivec], checksum_for(:tiger_g5)
-    assert_nil checksum_for(:tiger_g3)
   end
 end
