+++ Library/Homebrew/test/test_os_mac_bottle_tag.rb
@@ -1,7 +1,7 @@
 require "testing_env"
 require "utils/bottles"
 
-class BottleTagTests < Homebrew::TestCase
+class OSMacBottleTagTests < Homebrew::TestCase
   def test_tag_tiger_ppc
     MacOS.stubs(:version).returns(MacOS::Version.new("10.4"))
     Hardware::CPU.stubs(:type).returns(:ppc)
