+++ Library/Homebrew/utils/bottles.rb
@@ -5,7 +5,7 @@ module Utils
   class Bottles
     class << self
       def tag
-        @bottle_tag ||= "#{ENV["HOMEBREW_SYSTEM"]}-#{ENV["HOMEBREW_PROCESSOR"]}".downcase.to_sym
+        @bottle_tag ||= "#{ENV["HOMEBREW_SYSTEM"]}_#{ENV["HOMEBREW_PROCESSOR"]}".downcase.to_sym
       end
 
       def built_as?(f)
